AC-990: Added new billing end point
